Trait pcs::MultilinearCommitmentScheme

pub trait MultilinearCommitmentScheme<E: PairingEngine> {
    type ProverParam;
    type VerifierParam;
    type SRS;
    type Commitment;
    type Proof;
    type BatchProof;
    type Transcript;

    fn setup<R: RngCore>(
        rng: &mut R, 
        num_vars: usize
    ) -> Result<Self::SRS, PCSErrors>;
    fn commit(
        prover_param: &Self::ProverParam, 
        poly: &impl MultilinearExtension<E::Fr>
    ) -> Result<Self::Commitment, PCSErrors>;
    fn multi_commit(
        prover_param: &Self::ProverParam, 
        polys: &[impl MultilinearExtension<E::Fr>]
    ) -> Result<Self::Commitment, PCSErrors>;
    fn open(
        prover_param: &Self::ProverParam, 
        polynomial: &impl MultilinearExtension<E::Fr>, 
        point: &[E::Fr]
    ) -> Result<Self::Proof, PCSErrors>;
    fn multi_open(
        prover_param: &Self::ProverParam, 
        polynomials: &[impl MultilinearExtension<E::Fr>], 
        point: &[&[E::Fr]], 
        transcript: &mut Self::Transcript
    ) -> Result<Self::BatchProof, PCSErrors>;
    fn verify(
        verifier_param: &Self::VerifierParam, 
        commitment: &Self::Commitment, 
        point: &[E::Fr], 
        value: &E::Fr, 
        proof: &Self::Proof
    ) -> Result<bool, PCSErrors>;
    fn batch_verify(
        verifier_param: &Self::VerifierParam, 
        multi_commitment: &Self::Commitment, 
        points: &[&[E::Fr]], 
        batch_proof: &Self::BatchProof, 
        transcript: &mut IOPTranscript<E::Fr>
    ) -> Result<bool, PCSErrors>;
}

Required Associated Types

type ProverParam

type VerifierParam

type SRS

type Commitment

type Proof

type BatchProof

type Transcript

Required Methods

fn setup<R: RngCore>(
rng: &mut R,
num_vars: usize
) -> Result<Self::SRS, PCSErrors>

Generate SRS from RNG. WARNING: THIS FUNCTION IS FOR TESTING PURPOSE ONLY. THE OUTPUT SRS SHOULD NOT BE USED IN PRODUCTION.

fn commit(
prover_param: &Self::ProverParam,
poly: &impl MultilinearExtension<E::Fr>
) -> Result<Self::Commitment, PCSErrors>

Generate a commitment for a polynomial

fn multi_commit(
prover_param: &Self::ProverParam,
polys: &[impl MultilinearExtension<E::Fr>]
) -> Result<Self::Commitment, PCSErrors>

Generate a commitment for a list of polynomials

fn open(
prover_param: &Self::ProverParam,
polynomial: &impl MultilinearExtension<E::Fr>,
point: &[E::Fr]
) -> Result<Self::Proof, PCSErrors>

On input a polynomial p and a point point, outputs a proof for the same.

fn multi_open(
prover_param: &Self::ProverParam,
polynomials: &[impl MultilinearExtension<E::Fr>],
point: &[&[E::Fr]],
transcript: &mut Self::Transcript
) -> Result<Self::BatchProof, PCSErrors>

Input a list of MLEs, and a same number of points, and a transcript, compute a multi-opening for all the polynomials.

fn verify(
verifier_param: &Self::VerifierParam,
commitment: &Self::Commitment,
point: &[E::Fr],
value: &E::Fr,
proof: &Self::Proof
) -> Result<bool, PCSErrors>

Verifies that value is the evaluation at x of the polynomial committed inside comm.

fn batch_verify(
verifier_param: &Self::VerifierParam,
multi_commitment: &Self::Commitment,
points: &[&[E::Fr]],
batch_proof: &Self::BatchProof,
transcript: &mut IOPTranscript<E::Fr>
) -> Result<bool, PCSErrors>

Verifies that value_i is the evaluation at x_i of the polynomial poly_i committed inside comm.