Trait pcs::PolynomialCommitmentScheme

pub trait PolynomialCommitmentScheme<E: PairingEngine> {
    type ProverParam;
    type VerifierParam;
    type SRS;
    type Polynomial;
    type Point;
    type Evaluation;
    type Commitment: CanonicalSerialize;
    type BatchCommitment: CanonicalSerialize;
    type Proof;
    type BatchProof;

    fn gen_srs_for_testing<R: RngCore>(
        rng: &mut R, 
        log_size: usize
    ) -> Result<Self::SRS, PCSErrors>;
    fn trim(
        srs: &Self::SRS, 
        supported_log_degree: usize, 
        supported_num_vars: Option<usize>
    ) -> Result<(Self::ProverParam, Self::VerifierParam), PCSErrors>;
    fn commit(
        prover_param: &Self::ProverParam, 
        poly: &Self::Polynomial
    ) -> Result<Self::Commitment, PCSErrors>;
    fn multi_commit(
        prover_param: &Self::ProverParam, 
        polys: &[Self::Polynomial]
    ) -> Result<Self::BatchCommitment, PCSErrors>;
    fn open(
        prover_param: &Self::ProverParam, 
        polynomial: &Self::Polynomial, 
        point: &Self::Point
    ) -> Result<(Self::Proof, Self::Evaluation), PCSErrors>;
    fn multi_open(
        prover_param: &Self::ProverParam, 
        multi_commitment: &Self::Commitment, 
        polynomials: &[Self::Polynomial], 
        points: &[Self::Point]
    ) -> Result<(Self::BatchProof, Vec<Self::Evaluation>), PCSErrors>;
    fn verify(
        verifier_param: &Self::VerifierParam, 
        commitment: &Self::Commitment, 
        point: &Self::Point, 
        value: &E::Fr, 
        proof: &Self::Proof
    ) -> Result<bool, PCSErrors>;
    fn batch_verify<R: RngCore>(
        verifier_param: &Self::VerifierParam, 
        multi_commitment: &Self::BatchCommitment, 
        points: &[Self::Point], 
        values: &[E::Fr], 
        batch_proof: &Self::BatchProof, 
        rng: &mut R
    ) -> Result<bool, PCSErrors>;
}

This trait defines APIs for polynomial commitment schemes. Note that for our usage of PCS, we do not require the hiding property.

Required Associated Types

type ProverParam

type VerifierParam

type SRS

type Polynomial

type Point

type Evaluation

type Commitment: CanonicalSerialize

type BatchCommitment: CanonicalSerialize

type Proof

type BatchProof

Required Methods

fn gen_srs_for_testing<R: RngCore>(
rng: &mut R,
log_size: usize
) -> Result<Self::SRS, PCSErrors>

Build SRS for testing.

For univariate polynomials, log_size is the log of maximum degree.

For multilinear polynomials, log_size is the number of variables.

WARNING: THIS FUNCTION IS FOR TESTING PURPOSE ONLY. THE OUTPUT SRS SHOULD NOT BE USED IN PRODUCTION.

fn trim(
srs: &Self::SRS,
supported_log_degree: usize,
supported_num_vars: Option<usize>
) -> Result<(Self::ProverParam, Self::VerifierParam), PCSErrors>

Trim the universal parameters to specialize the public parameters. Input both supported_log_degree for univariate and supported_num_vars for multilinear.

fn commit(
prover_param: &Self::ProverParam,
poly: &Self::Polynomial
) -> Result<Self::Commitment, PCSErrors>

Generate a commitment for a polynomial

fn multi_commit(
prover_param: &Self::ProverParam,
polys: &[Self::Polynomial]
) -> Result<Self::BatchCommitment, PCSErrors>

Generate a commitment for a list of polynomials

fn open(
prover_param: &Self::ProverParam,
polynomial: &Self::Polynomial,
point: &Self::Point
) -> Result<(Self::Proof, Self::Evaluation), PCSErrors>

On input a polynomial p and a point point, outputs a proof for the same.

fn multi_open(
prover_param: &Self::ProverParam,
multi_commitment: &Self::Commitment,
polynomials: &[Self::Polynomial],
points: &[Self::Point]
) -> Result<(Self::BatchProof, Vec<Self::Evaluation>), PCSErrors>

Input a list of MLEs, and a same number of points, and a transcript, compute a multi-opening for all the polynomials.

fn verify(
verifier_param: &Self::VerifierParam,
commitment: &Self::Commitment,
point: &Self::Point,
value: &E::Fr,
proof: &Self::Proof
) -> Result<bool, PCSErrors>

Verifies that value is the evaluation at x of the polynomial committed inside comm.

fn batch_verify<R: RngCore>(
verifier_param: &Self::VerifierParam,
multi_commitment: &Self::BatchCommitment,
points: &[Self::Point],
values: &[E::Fr],
batch_proof: &Self::BatchProof,
rng: &mut R
) -> Result<bool, PCSErrors>

Verifies that value_i is the evaluation at x_i of the polynomial poly_i committed inside comm.